fix: 목데이터 수정
